Extract scraper into a class
This will allow us to add test for the methods that were already present
and to refactor further.
